* Fix dependency. * Remove the legacy project.
