+++ ScintillaTabbedTextControl.NET/VPKSoft.ScintillaTabbedTextControl.NET.csproj
@@ -20,19 +20,6 @@
 		<PackageReleaseNotes>See: https://github.com/VPKSoft/ScintillaTabbedTextControl</PackageReleaseNotes>
 		<UseWindowsForms>true</UseWindowsForms>
 	</PropertyGroup>
-	<ItemGroup>
-		<Compile Include="..\ScintillaTabbedTextControl\FileTabButton.cs" Link="FileTabButton.cs" />
-		<Compile Include="..\ScintillaTabbedTextControl\FileTabButton.Designer.cs" Link="FileTabButton.Designer.cs" />
-		<Compile Include="..\ScintillaTabbedTextControl\ScintillaTabbedDocument.cs" Link="ScintillaTabbedDocument.cs" />
-		<Compile Include="..\ScintillaTabbedTextControl\ScintillaTabbedTextControl.cs" Link="ScintillaTabbedTextControl.cs" />
-		<Compile Include="..\ScintillaTabbedTextControl\ScintillaTabbedTextControl.Designer.cs" Link="ScintillaTabbedTextControl.Designer.cs" />
-		<Compile Include="..\ScintillaTabbedTextControl\ScintillaZoomPercentage.cs" Link="ScintillaZoomPercentage.cs" />
-		<Compile Include="..\ScintillaTabbedTextControl\WheelScrollButtons.cs" Link="WheelScrollButtons.cs" />
-	</ItemGroup>
-	<ItemGroup>
-		<EmbeddedResource Include="..\ScintillaTabbedTextControl\FileTabButton.resx" Link="FileTabButton.resx" />
-		<EmbeddedResource Include="..\ScintillaTabbedTextControl\ScintillaTabbedTextControl.resx" Link="ScintillaTabbedTextControl.resx" />
-	</ItemGroup>
 	<ItemGroup Condition=" '$(TargetFramework)' == 'net47' ">
 		<Reference Include="System.Windows.Forms" />
 		<Reference Include="System.Data.DataSetExtensions" />
@@ -43,12 +30,6 @@
 		<PackageReference Include="VPKSoft.ScintillaLexers.NET" Version="1.1.15" />
 	</ItemGroup>
 	<ItemGroup>
-		<Compile Update="NoFocusButton.cs">
-			<SubType>Component</SubType>
-		</Compile>
-		<Compile Update="NoFocusButton.Designer.cs">
-			<DependentUpon>NoFocusButton.cs</DependentUpon>
-		</Compile>
 		<Compile Update="Properties\Resources.Designer.cs">
 			<DesignTime>True</DesignTime>
 			<AutoGen>True</AutoGen>
