--- ScintillaTabbedTextControl/VPKSoft.ScintillaTabbedTextControl.csproj
@@ -1,97 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-  <PropertyGroup>
-    <TargetFramework>net47</TargetFramework>
-    <AssemblyTitle>VPKSoft.ScintillaTabbedTextControl</AssemblyTitle>
-    <Company>VPKSoft</Company>
-    <Product>VPKSoft.ScintillaTabbedTextControl</Product>
-    <Description>A tabbed control for ScintillaNET to display multiple documents.</Description>
-    <Copyright>Copyright © VPKSoft 2021</Copyright>
-    <DocumentationFile>bin\$(Configuration)\VPKSoft.ScintillaTabbedTextControl.xml</DocumentationFile>
-    <OutputPath>bin\$(Configuration)\</OutputPath>
-    <Authors>VPKSoft</Authors>
-    <Version>1.1.5</Version>
-    <PackageLicenseExpression>MIT</PackageLicenseExpression>
-    <PackageProjectUrl>https://github.com/VPKSoft/ScintillaTabbedTextControl</PackageProjectUrl>
-    <PackageIcon>vpksoft_logo.png</PackageIcon>
-    <RepositoryUrl>https://github.com/VPKSoft/ScintillaTabbedTextControl</RepositoryUrl>
-    <RepositoryType>git</RepositoryType>
-    <PackageTags>wrapper scintilla ScintillaNET winforms control tabpage tab</PackageTags>
-    <PackageReleaseNotes>See: https://github.com/VPKSoft/ScintillaTabbedTextControl</PackageReleaseNotes>
-    <GeneratePackageOnBuild>true</GeneratePackageOnBuild>
-    <PackageRequireLicenseAcceptance>true</PackageRequireLicenseAcceptance>
-    <UserSecretsId>eb070337-a19d-407a-87cb-90420745c88e</UserSecretsId>
-  </PropertyGroup>
-  <PropertyGroup Condition="'$(Configuration)|$(TargetFramework)|$(Platform)'=='Debug|netcoreapp3.1|AnyCPU'">
-    <PlatformTarget>AnyCPU</PlatformTarget>
-  </PropertyGroup>
-  <PropertyGroup Condition="'$(Configuration)|$(TargetFramework)|$(Platform)'=='Debug|net47|AnyCPU'">
-    <PlatformTarget>AnyCPU</PlatformTarget>
-  </PropertyGroup>
-  <ItemGroup>
-    <Reference Include="System.Data.DataSetExtensions" />
-    <Reference Include="Microsoft.CSharp" />
-    <Reference Include="System.Net.Http" />
-    <Reference Include="System.Windows.Forms" />
-  </ItemGroup>
-  <ItemGroup>
-    <Compile Update="FileTabButton.cs">
-      <SubType>UserControl</SubType>
-    </Compile>
-    <Compile Update="FileTabButton.Designer.cs">
-      <DependentUpon>FileTabButton.cs</DependentUpon>
-    </Compile>
-    <Compile Update="NoFocusButton.cs">
-      <SubType>Component</SubType>
-    </Compile>
-    <Compile Update="NoFocusButton.Designer.cs">
-      <DependentUpon>NoFocusButton.cs</DependentUpon>
-    </Compile>
-    <Compile Update="Properties\Resources.Designer.cs">
-      <AutoGen>True</AutoGen>
-      <DesignTime>True</DesignTime>
-      <DependentUpon>Resources.resx</DependentUpon>
-    </Compile>
-    <Compile Update="ScintillaTabbedTextControl.cs">
-      <SubType>UserControl</SubType>
-    </Compile>
-    <Compile Update="ScintillaTabbedTextControl.Designer.cs">
-      <DependentUpon>ScintillaTabbedTextControl.cs</DependentUpon>
-    </Compile>
-  </ItemGroup>
-  <ItemGroup>
-    <EmbeddedResource Update="FileTabButton.resx">
-      <DependentUpon>FileTabButton.cs</DependentUpon>
-    </EmbeddedResource>
-    <EmbeddedResource Update="ScintillaTabbedTextControl.resx">
-      <DependentUpon>ScintillaTabbedTextControl.cs</DependentUpon>
-      <SubType>Designer</SubType>
-    </EmbeddedResource>
-  </ItemGroup>
-  <ItemGroup>
-    <None Include="..\vpksoft_logo.png">
-      <Pack>True</Pack>
-      <PackagePath></PackagePath>
-    </None>
-    <None Include="Resources\Playback.png" />
-    <None Include="ScintillaTabbedTextControl.licenseheader" />
-  </ItemGroup>
-  <ItemGroup>
-    <None Include="Resources\Play.png" />
-  </ItemGroup>
-  <ItemGroup>
-    <None Include="Resources\Save.png" />
-  </ItemGroup>
-  <ItemGroup>
-    <None Include="Resources\Save_Red.png" />
-  </ItemGroup>
-  <ItemGroup>
-    <None Include="Resources\Cancel.png" />
-  </ItemGroup>
-  <ItemGroup>
-    <None Include="Resources\New document.png" />
-  </ItemGroup>
-  <ItemGroup>
-    <PackageReference Include="jacobslusser.ScintillaNET" Version="3.6.3" />
-    <PackageReference Include="VPKSoft.ScintillaLexers" Version="1.1.7" />
-  </ItemGroup>
-</Project>
