+++ VPKSoft.ScintillaTabbedTextControl.sln
@@ -5,8 +5,6 @@ VisualStudioVersion = 17.0.32014.148
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TestApp", "TestApp\TestApp.csproj", "{33F3701A-FEFE-4BD9-822C-55F80541E797}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "VPKSoft.ScintillaTabbedTextControl", "ScintillaTabbedTextControl\VPKSoft.ScintillaTabbedTextControl.csproj", "{7522F687-CE76-4AF0-B3CD-9C027EDB5D5D}"
-EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{8F0F61DA-9D7A-4A6B-9630-9411A1B5DE24}"
 	ProjectSection(SolutionItems) = preProject
 		.github\workflows\dotnet-desktop.yml = .github\workflows\dotnet-desktop.yml
@@ -24,10 +22,6 @@ Global
 		{33F3701A-FEFE-4BD9-822C-55F80541E797}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{33F3701A-FEFE-4BD9-822C-55F80541E797}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{33F3701A-FEFE-4BD9-822C-55F80541E797}.Release|Any CPU.Build.0 = Release|Any CPU
-		{7522F687-CE76-4AF0-B3CD-9C027EDB5D5D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{7522F687-CE76-4AF0-B3CD-9C027EDB5D5D}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{7522F687-CE76-4AF0-B3CD-9C027EDB5D5D}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{7522F687-CE76-4AF0-B3CD-9C027EDB5D5D}.Release|Any CPU.Build.0 = Release|Any CPU
 		{A9FD52F8-65AD-4EF5-9C5A-9F76A487FA73}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{A9FD52F8-65AD-4EF5-9C5A-9F76A487FA73}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{A9FD52F8-65AD-4EF5-9C5A-9F76A487FA73}.Release|Any CPU.ActiveCfg = Release|Any CPU
